refactor: extract 35 UI primitive methods to ui_primitives.py
Move all low-level UI interaction methods (element finding, clicking,
text reading, coordinate utilities) from DamaiBot into a UIPrimitives
mixin class. DamaiBot now inherits from UIPrimitives, preserving full
backward compatibility. damai_app.py reduced from 3600 to 2959 lines.
